requirements.txt
langchain==0.0.168
matplotlib==3.7.1
python-dotenv==1.0.0
steamship==2.16.9
